Adds new macro to specify enum configurations with custom option name strings
This commit introduces a new macro definition called
`enum_configuration2` to help with declaring a enum configuration type
where the developer can specify a custom option name string in the enum
option value.

Example:
```rust
enum_configuration2! {
    #[derive(PartialEq)]
    enum EnumConfiguration2 {
        Val1 = ("val_1", 1),
        Val2 = ("val_2", 2),
    }
}
```

The enum option value string accepted by the `CONFIG SET` command will be
`val_1` or `val_2` instead of `Val1` or `Val2` as in when declaring the
enum type with the `enum_configuration` macro.

This commit also fixes a bug in the `enum_configuration` macro, which
would not allow the macro to be used twice in the same file.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>